# backend/src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"RoleDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRespDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/RoleDTO"
						},
						"type": "array"
					},
					"personalDetails": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserReqDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"roleIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"status": {
						"type": "string"
					},
					"personalDetails": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"roleIds"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ResourceDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"location": {
						"type": "string"
					},
					"sourceId": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"group": {
						"type": "string"
					}
				},
				"required": [
					"location",
					"type",
					"group"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EntityDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"regId": {
						"type": "string"
					},
					"ownerId": {
						"type": "string"
					},
					"caId": {
						"type": "string"
					},
					"settings": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"name",
					"address",
					"country",
					"regId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectRespDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"countryName": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"actualMaturityDate": {
						"type": "string",
						"format": "date-time"
					},
					"overallCost": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"ownerName": {
						"type": "string"
					},
					"legalId": {
						"type": "string"
					},
					"maturityLockingPeriod": {
						"type": "number",
						"format": "double"
					},
					"settings": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"entity": {
						"$ref": "#/components/schemas/EntityDTO"
					},
					"resources": {
						"items": {
							"$ref": "#/components/schemas/ResourceDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"address",
					"countryName",
					"startDate",
					"entity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectReqDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"latitude": {
						"type": "number",
						"format": "double"
					},
					"countryName": {
						"type": "string"
					},
					"longitude": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string"
					},
					"actualMaturityDate": {
						"type": "string"
					},
					"overallCost": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"ownerName": {
						"type": "string"
					},
					"legalId": {
						"type": "string"
					},
					"maturityLockingPeriod": {
						"type": "number",
						"format": "double"
					},
					"settings": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"entityID": {
						"type": "number",
						"format": "double"
					},
					"resourceIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"address",
					"countryName",
					"startDate",
					"entityID"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "nexgen-services",
		"version": "1.0.0",
		"description": "**Edit a file, create a new file, and clone from Bitbucket in under 2 minutes**",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserRespDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserReqDTO"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserReqDTO"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{id}/roles": {
			"put": {
				"operationId": "UpdateUserRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/projects": {
			"get": {
				"operationId": "GetAllProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProjectRespDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ProjectRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectReqDTO"
							}
						}
					}
				}
			}
		},
		"/projects/{id}": {
			"get": {
				"operationId": "GetProjectById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ProjectRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ProjectRespDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"settings": {
										"properties": {},
										"additionalProperties": {},
										"type": "object"
									},
									"ownerName": {
										"type": "string"
									},
									"maturityLockingPeriod": {
										"type": "number",
										"format": "double"
									},
									"overallCost": {
										"type": "number",
										"format": "double"
									},
									"description": {
										"type": "string"
									},
									"address": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}